Don't call save() from move().
I want to call move() in several situations, including validating whether a
move will put or leave a king in check, and during automated testing. I only
want to save the state of the board in one situation: after a real move in the
game, as opposed to hypothetical moves. Therefore, it makes sense to call save
explicitly.