Fix updateDrawCount().
Although it was resetting the count to zero, it was still counting pawn &
capture moves, so it was calling 75-move draws too early.

Also, it wasn't counting the move when a draw had already been detected, which
isn't a huge deal, but did mean that the draw count would be too low by one if
the game ended in another type of draw. It might be interesting to know if a
draw occurs in more than one way simultaneously.