Memoize analyze().
I chose to store the move data in a static property so that there is a common
pool for all analysis, including the work of validateMove() and evaluateMove().
I also like that this data doesn't get saved: I don't want it to grow too
large, or preclude future changes from taking effect.