Simplify emulating moves.
The bot evaluates a move two ways: based on the current state of the board, and
by emulating the move and seeing the effects.

Previously, the bot emulated each move twice: once to see if the move results
in checkmate, and if not, again to see if the move results in a draw or check.

Now, the emulation path assigns a meaningful rating and the caller simply adds
this value to the move's rating. This simplifies the calling code, since it no
longer needs to handle the results of emulation differently than any other
rating.