Count trapped pieces.
Consider whether each move affects how many pieces can move.

At the beginning of the game, 6 pieces are trapped: both rooks, both bishops,
the queen and the king. Arguably, moving the e-file pawn is most wise, since
that allows the queen and the kingside bishop to develop, and offers the king
an escape route should a knight bring check.