Align piece symbols on the bottom.
This is an attempt to fix a layout bug on iOS. Squares will resize vertically
depending on which pieces are in them. I noticed that for the taken pieces, the
black queen, black rooks, and black knights are not aligned along the bottom of
their containers the way the other pieces are.